+++ app/src/main/java/com/nixstudio/moviemax/models/sources/remote/DiscoverTvResponse.kt
@@ -57,8 +57,8 @@ data class DiscoverTvResultsItem(
 	val name: String? = null,
 
 	@field:SerializedName("id")
-	val id: Int? = null,
+	val id: Long? = null,
 
 	@field:SerializedName("vote_count")
-	val voteCount: Int? = null
+	val voteCount: Long? = null
 ): Parcelable
